pyproject.toml
[tool.poetry]
name = "dpost-backend"
version = "0.1.0"
description = ""
authors = ["jw.park"]
license = "MIT"
readme = "README.md"
packages = [{include = "dpost_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
